requirements.txt
nltk==3.6.3
spotipy==2.19.0
python-dotenv==0.19.1